DevFreela.Application/Commands/LoginUser/LoginUserCommandHandler.cs
using DevFreela.Application.ViewModels;
using DevFreela.Core.Repositories;
using DevFreela.Core.Service;
using MediatR;
using System.Threading;
using System.Threading.Tasks;

namespace DevFreela.Application.Commands.LoginUser
{
    public class LoginUserCommandHandler : IRequestHandler<LoginUserCommand, LoginUserViewModel>
    {
        private readonly IAuthService _authService;
        private readonly IUserRepository _userRepository;

        public LoginUserCommandHandler(IAuthService authService, IUserRepository userRepository)
        {
            _authService = authService;
            _userRepository = userRepository;
        }

        public async Task<LoginUserViewModel> Handle(LoginUserCommand command, CancellationToken cancellationToken)
        {
            var passwordHash = _authService.ComputeSha256Hash(command.Password);

            var user = await _userRepository.GetByEmailAndPasswordAsync(command.Email, passwordHash);

            if (user == null) return null;

            var token = _authService.GenerateJwtToken(user.Email, user.UserRoles);
            return new LoginUserViewModel(user.Email, token);
        }
    }
}
